requirements.txt
django
djangorestframework
requests
pep8
pillow
django-cors-headers
djangorestframework-xml
